// projects/binom/sdk-layout/css-grid/src/interfaces/bn-grid-cur-states.ts
export interface BnGridCurStates {
    isServiceEvent:  boolean;
    fullScreenEvent:  boolean;
    fullScreenState:  boolean;
    iconsSidebarEvent:  boolean;
    iconsSidebarState:  boolean;
    fixedChanged:  boolean;
    isFixed: boolean;
    resizeEvent:  boolean;
    visibleChanged:  boolean;
    sideBarVisibleLeftState:  boolean;
    sideBarVisibleRightState: boolean;
    lastSideBarSource:string;
}